[Refactor]: Moved projection2image-type viz functions to a new file (#480)
